Add salary methods, update data types, enhance charts
Added AddSalaryDetailSummariesAsync and UpdateSalaryDetailSummariesAsync methods to SalaryDetailRepository and ISalaryDetailRepository. Changed data types of salary-related fields from decimal to int. Updated ISalaryService and SalaryService to reflect these changes. Modified service registrations in ApplicationServiceExtensions. Enhanced ExpenditurePageViewModel with LiveCharts for pie chart visualization and updated ExpenditurePage.xaml accordingly.

+++ SandTetris/Entities/SalaryDetail.cs
@@ -6,9 +6,9 @@ public class SalaryDetail
     public Employee Employee { get; set; } = null!;
     public int Month { get; set; }
     public int Year { get; set; }
-    public decimal BaseSalary { get; set; }
+    public int BaseSalary { get; set; }
 
     public int DaysAbsent { get; set; }
     public int DaysOnLeave { get; set; }
-    public decimal FinalSalary { get; set; }
+    public int FinalSalary { get; set; }
 }
